SCA cleanup

+++ STM32 FW Project/Official release with BLE Remocon - 170318/Src/basic_math.c
@@ -9,7 +9,7 @@ float32_t Sqrt(float32_t x)
     float32_t x;
   } u;
   u.x = x;
-  u.i = ((1<<29) + (u.i >> 1)) - (1<<22);
+  u.i = ((1 << 29) + (u.i >> 1)) - (1 << 22);
   
   // Two Babylonian Steps (simplified from:)
   // u.x = 0.5f * (u.x + x/u.x);
@@ -29,9 +29,9 @@ float32_t invSqrt(float32_t x)
 {
 	float32_t halfx = 0.5f * x;
 	float32_t y = x;
-	long i = *(long*)&y;
-	i = 0x5f3759df - (i>>1);
-	y = *(float32_t*)&i;
+	int32_t i = *(int32_t*) &y;
+	i = 0x5f3759df - (i >> 1);
+	y = *(float32_t*) &i;
 	y = y * (1.5f - (halfx * y * y));
 	return y;
 }
